split ffi.rs into core/ffi.rs and lib.rs
